ci: prep contracts for maintenance
