+++ contracts/child/validator/legacy-compat/@openzeppelin/contracts-upgradeable/README.md
@@ -1,5 +1,9 @@
-## Changes
+### Changes
 
-- The `__gap`s have been trimmed down in all contracts.
-- `LegacyStateVars` have been injected into `EIP712Upgradeable`.
-- The import paths have been adjusted.
+- The `__gap`s have been trimmed down in all contracts
+- `LegacyStateVars` have been injected into `EIP712Upgradeable`
+- The import paths have been adjusted
+
+### Checks
+
+See the [script](../../../../../../test/forge/hardfork/utils/checkModifiedOpenZeppelinContracts.js)
